feat: let the build script could install dependences.

+++ buildkernel.sh
@@ -3,8 +3,40 @@ export CC=clang
 export PATH=$PATH:$(pwd)/build-shit/clang/bin
 export CROSS_COMPILE_ARM32=$(pwd)/build-shit/gcc-arm/bin/arm-eabi-
 export CROSS_COMPILE=$(pwd)/build-shit/gcc-arm64/bin/aarch64-elf-
+export CLANG_URL="https://android.googlesource.com/platform//prebuilts/clang/host/linux-x86/+archive/3857008389202edac32d57008bb8c99d2c957f9d/clang-r383902.tar.gz"
 
-rm -rf ./out
+setup_environment() {
+    sudo apt install -y lib32ncurses5-dev automake libssl-dev lzop policycoreutils xsltproc libbz2-1.0 \
+                        lib32z-dev unzip bzip2 maven liblz4-tool libghc-bzlib-dev libxml-simple-perl \
+                        build-essential flex gperf pngcrush libxml-sax-base-perl x11proto-core-dev minicom \
+                        libswitch-perl ccache libxml2-utils git bison pwgen zip schedtool wget \
+                        libx11-dev bc libc6-dev-i386 g++-multilib zlib1g-dev dpkg-dev libelf-dev libdebuginfod-dev \
+                        squashfs-tools optipng curl libgl1-mesa-dev libbz2-dev make gnupg libncurses5 fontconfig
+    mkdir -p build-shit && cd build-shit || exit
+    git clone https://github.com/KenHV/gcc-arm64.git --single-branch -b master --depth=1 || exit
+    git clone https://github.com/KenHV/gcc-arm.git --single-branch -b master --depth=1 || exit
+    mkdir clang && cd clang || exit
+    wget $CLANG_URL -o clang.tar.gz || exit
+    tar -xzvf clang.tar.gz || exit
+    export PATH=$(pwd):$PATH
+    cd ../..
+}
+
+if [! -d "build-shit"]
+then
+    setup_environment
+fi
+
+if [ -d "out"]
+then
+    rm -rf ./out
+fi
+
+if [ -d "CosmicFresh/Image"]
+then
+    rm CosmicFresh/Image -f
+fi
+
 make O=out -j$(nproc --all) nabu_defconfig || exit
 make O=out -j$(nproc --all) || exit
